pyproject.toml
[tool.poetry]
name = "showcase-currency"
version = "0.1.0"
description = ""
authors = ["Dmitry Gening <geningdm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.5"
aioredis = "^1.2"
marshmallow = "^2.19"
pytest-aiohttp = "^0.3.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-aiohttp = "^0.3.0"
factory-boy = "^2.12"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
